updated register functionality
