+++ src/repositories/user-repository.js
@@ -7,7 +7,18 @@ const getUserByEmail = async (email) =>{
 }
 
 
+const saveUser = async (newUser) => {
+
+    // if newUser id is not null, it will update the user. If  id is null it will create a new register
+
+    const user = User.build(newUser,{isNewRecord: !newUser.id})
+
+    await user.save()
+
+    return user
+}
 
 module.exports = {
-    getUserByEmail
+    getUserByEmail,
+    saveUser
 }
