+++ src/routes/auth.js
@@ -1,10 +1,12 @@
 const express = require("express");
 const router = express.Router();
 const authController = require('../controllers/auth-controller')
-const authService = require ('../services/auth-service')
 
+
+const ValidationOfTheRegistrationForm = require('../middlewares/validations/ValidationOfTheRegistrationForm')
 
 router.post("/login", authController.login)
 
+router.post("/register", ValidationOfTheRegistrationForm, authController.register)
 
 module.exports = router
