+++ src/routes/users.js
@@ -1,7 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const userController = require('../controllers/users-controllers');
-const ValidationOfTheRegistrationForm = require("../middlewares/validations/ValidationOfTheRegistrationForm");
+
